Sequential quality score calculation is now in place in the refactored recalibrator and matches the quality scores calculated by the old recalibrator exactly; at least on the small sets of data used so far. Validation, documentation, and optimization work is on going.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@1985 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/playground/gatk/walkers/Recalibration/MappingQualityCovariate.java
@@ -13,7 +13,7 @@ public class MappingQualityCovariate implements Covariate {
     }
 
     public Comparable<?> getValue(SAMRecord read, int offset, char[] refBases) {
-        return read.getMappingQuality();
+        return (Integer)(read.getMappingQuality());
     }
     
     public Comparable<?> getValue(String str) {
